pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 42", "versioningit ~= 1.0", "wheel"]

[project]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Topic :: Other/Nonlisted Topic",
    "Topic :: Scientific/Engineering",
]
authors = [{name = "Jimmy-Xuan Shen", email = "jmmshn@gmail.com"}]
dependencies = [
  'pymatgen>=2022.3.22',
  "h5py>=2.10.0",
]
description = "Tools for re-griding periodic volumetric quantum chemistry data for machine-learning purposes."
dynamic = ["version"]
keywords = ["machine-learning", "dft", "vasp", "volumetric", "pymatgen"]
license = {text = "modified BSD"}
name = "mp-pyrho"
readme = "README.md"
requires-python = '>="3.8"'

[project.optional-dependencies]
dev = ["pre-commit==2.17.0"]
docs = [
  "numpydoc==1.3.1",
]
tests = [
  "pytest==7.1.2",
  "pytest-cov==3.0.0",
  "hypothesis==6.40.1",
]
strict = [
  'pymatgen==2022.4.26',
  "h5py==3.6.0",
]

[project.urls]
homepage = "https://materialsproject.github.io/pyrho/"
repository = "https://materialsproject.github.io/pyrho"

[tool.versioningit.vcs]
default-tag = "0.0.1"
method = "git"

[tool.isort]
profile = "black"

[tool.flake8]
docstring-convention = "numpy"
extend-ignore = "E203, W503, E501, F401, RST21"
max-doc-length = 120
max-line-length = 120
min-python-version = "3.8.0"
rst-roles = "class, func, ref, obj"
select = "C, E, F, W, B, B950"

[tool.mypy]
ignore_missing_imports = true
no_strict_optional = true

[tool.pytest.ini_options]
filterwarnings = [
  "ignore:.*POTCAR.*:UserWarning",
  "ignore:.*magmom.*:UserWarning",
  "ignore:.*is not gzipped.*:UserWarning",
  "ignore:.*input structure.*:UserWarning",
  "ignore::DeprecationWarning",
]

[tool.coverage.run]
branch = true
include = ["src/*"]
parallel = true

[tool.coverage.paths]
source = ["src/"]

[tool.coverage.report]
exclude_lines = [
  '^\s*assert False(,|$)',
  'if typing.TYPE_CHECKING:',
  '^\s*@overload( |$)',
]
show_missing = true
skip_covered = true
